project/requirements.txt
appdirs==1.4.4
asgiref==3.7.2
asttokens==2.2.1
attrs==23.1.0
Babel==2.10.3
backcall==0.2.0
bcrypt==3.2.2
beautifulsoup4==4.12.2
beniget==0.4.1
bitcoin-value==1.6.0
black==23.12.1
blinker==1.6.2
breezy==3.3.4
Brlapi==0.8.4
Brotli==1.0.9
bs4==0.0.1
certifi==2022.9.24
chardet==5.1.0
chromedriver-autoinstaller==0.6.4
click==8.1.6
cloud-init==23.4.4
colorama==0.4.6
command-not-found==0.3
configobj==5.0.8
contourpy==1.0.7
cryptocompare==0.7.6
cryptography==38.0.4
cups-of-caffeine==2.9.12
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.3.2
decorator==5.1.1
defer==1.0.6
Deprecated==1.2.14
distlib==0.3.8
distro==1.8.0
distro-info==1.5+ubuntu0.23.10.1
Django==5.0.1
django-next-prev==1.1.0
dulwich==0.21.6
duplicity==1.2.2
ewmh==0.1.6
executing==1.2.0
fastbencode==0.2
fasteners==0.17.3
fastimport==0.9.14
filelock==3.13.1
fonttools==4.38.0
fs==2.4.16
future==0.18.2
gast==0.5.2
gpg==1.18.0
h11==0.14.0
html5lib==1.1
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
ipython==8.14.0
jaraco.classes==3.2.1
jedi==0.18.2
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.10.3
keyring==24.2.0
kiwisolver==0.0.0
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
led==0.0.2
lockfile==0.12.2
louis==3.26.0
lxml==4.9.3
lz4==4.0.2+dfsg
Mako==1.2.4.dev0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdurl==0.1.2
merge3==0.0.8
monotonic==1.6
more-itertools==10.1.0
mpmath==0.0.0
mypy-extensions==1.0.0
mysql-connector-python==8.2.0
mysqlclient==2.2.4
netifaces==0.11.0
numpy==1.24.2
oauthlib==3.2.2
olefile==0.46
outcome==1.3.0.post0
packaging==23.1
pandas==2.2.1
paramiko==2.12.0
parso==0.8.3
pathspec==0.12.1
patiencediff==0.2.13
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==4.1.0
playsound==1.3.0
ply==3.11
prompt-toolkit==3.0.39
protobuf==4.21.12
ptyprocess==0.7.0
pure-eval==0.0.0
pycairo==1.24.0
pycups==2.0.1
pygame==2.5.2
PyGithub==1.55
Pygments==2.15.1
PyGObject==3.46.0
PyJWT==2.7.0
PyNaCl==1.5.0
pyparsing==3.1.0
pyrsistent==0.18.1
pyserial==3.5
PySocks==1.7.1
pysqlite3-binary==0.5.2.post3
python-apt==2.6.0+ubuntu1
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-dotenv==1.0.1
python-xlib==0.33
pythran==0.11.0
pyttsx3==2.90
pytz==2023.3
pyxdg==0.28
PyYAML==6.0.1
requests==2.31.0
rich==13.3.1
scikit-learn==1.4.1.post1
scipy==1.10.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.3
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
sympy==1.12
systemd-python==235
threadpoolctl==3.3.0
traitlets==5.5.0
trio==0.23.2
trio-websocket==0.11.1
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.14.0
ufw==0.36.2
unattended-upgrades==0.1
urllib3==1.26.16
usb-creator==0.3.16
virtualenv==20.25.0
wadllib==1.3.6
wcwidth==0.2.5
webdriver-manager==4.0.1
webencodings==0.5.1
wrapt==1.14.1
wsproto==1.2.0
xdg==5
xkit==0.0.0
zipp==1.0.0
